refactor: extract flatten operations into wvImportFlattenController
Move syncInstances and handleAsGrouping out of the
worldViewImportController monolith into a dedicated
wvImportFlattenController file. handleAsGrouping reads/writes
undoEntries from wvImportSharedState (introduced in PR-14).

Pure refactor — no behavior change, byte-identical function bodies.
Continues the spine domain-split (ADR-0009).
